build: ping python and upgrade poetry (#8)
* Fix Oracle build

And directly download instantclient.

* Add INSTANTCLIENT_VDIR

* Update Snowflake client

* Update copyright

* Upgrade Exasol clients

And show cpan module build logs on failure.

* Add Git SHA tag to images

* Upgrade to v1.2.0

* Push all tags

And remove mental health term from the descripion

* Push only on tag

* Restore latest tag, add git tag

* Upgrade Sqitch to v1.2.1

* Upgrade Oracle, Snowflake, Exasol clients

* Fix flags when running without TTY

Make sure the container can be run without a TTY, e.g. deploying changes
when testing in a CI pipeline.

The following pipes and redirects are now possible:

    $        ./docker-sqitch.sh
    $        ./docker-sqitch.sh < file
    $        ./docker-sqitch.sh < file | less
    $        ./docker-sqitch.sh        | less
    $ echo | ./docker-sqitch.sh        | less
    $ echo | ./docker-sqitch.sh
    $ echo | ./docker-sqitch.sh > file
    $        ./docker-sqitch.sh > file

* Upgrade to Sqitch v1.3.0

* Prepend v to version tags and labels

* Document Cockroach and Yugabyte support

* Upgrade Sqitch to v1.3.1

* Use checkout@v3

* Fix Snowflake config file permission issue

And upgrade the ODBC driver. Thanks to @peter-wimsey
for the fix! Closes #49.

* Try docker login action

* Update README.md

Fix file creation for Windows

* Upgrade drivers and CLIs

* Skip testing, at least for now

* Upgarde the Snowflake, Oracle, & Exasol clients

And keep the build on Bullseye for now. Upgrade Open JDK for the Exasol
image since v11 is no longer available.

* Upgrade Sqitch to v1.4.0

* Remove wayward Data::Dump

* Upgrade Snowflake and Oracle clients

And add the misssing SNOWSQL_ROLE environment variable to
docker-sqitch.sh.

* Build images for ARM64

Update the CI/CD workflow to use the latest Docker actions for
assembling metadata, building, testing, and pushing multi-platform
Docker images. Cease using the `./build` script in the workflow, but
update it to build images for multiple platforms. To have both the CI/CD
workflow and the `./build` script in sync, add the `.envrc` file and
teach both to read from it for common values, especially the Sqitch
version.

Fix an issue building Sqitch on ARM64 by installing List::MoreUtils::XS
before building Sqitch; otherwise it hangs and times out during the
configure phase for that module.

Update the Oracle `Dockerfile` to always install the latest Instant
Client; no more need to edit it for new versions. Also teach it to
properly install ARM64 client libraries, but disable building for ARM64
in the `Makefile`, for now, until a [bug building DBD::Oracle] can be
fixed.

Update the Snowflake `Dockerfile` to install the ARM64 ODBC driver on
ARM64 Linux, but limit builds to AMD64 in the `Makefile`, for now, until
[SnowSQL gets ARM64 support].

Limit the Exasol build to AMD64, as there are currently no Exasol
libraries for ARM64.

Remove reference to Vertica, for which there is not yet Docker support.

  [bug building DBD::Oracle]: https://rt.cpan.org/Ticket/Display.html?id=149876
  [SnowSQL gets ARM64 support]: https://community.snowflake.com/s/question/0D5Do00000ltxpVKAQ/snowsql-linuxarm64-support

* Switch from curl to WebRequest on Windows

Closes #60.

* Upgrade Sqitch to v1.4.1

* Fix bundle bug, upgrade to Bookworm

* build: ping python and upgrade poetry

---------

Co-authored-by: David E. Wheeler <david@justatheory.com>
Co-authored-by: Erik Wienhold <ewie@ewie.name>
Co-authored-by: Ruan Heyns <58415372+RHeynsZa@users.noreply.github.com>